improvement(docs): do not use `build` field in migration guide (#6207)
During the Bonsai release there was some confusion around the best
practices for referencing a container build in a `container` Deploy
action.

Originally we used the `build` field for this but it has caused
confusion and only works for Deploy actions of type `container`, not
e.g. `helm` or `kubernetes`.

So for consistency, we decided that the best practice should be to
always explicitly set the corresponding Build action as a dependency for
ALL Deploy actions (i.e. any type).

This commit updates the Bonsai migration guide to reflect that.
Previously we've updated all our examples and other parts of our
documentation.