chore: ensure consistency of `ContextNode` keys and add new tests (#6254)
* refactor: a bit more verbose code for traversing tree nodes

To make it a bit easier to read and debug.

* chore: add JSDocs for `ContextNode` to explain its semantics

* test: add test for the current semantics of `getByContext`

* test: add test to ensure internal state and invariants of `TreeCache`

* fix: prevent unwanted key mutations in `ContextNode`

Use a copy of a context key array when creating a new node.

Otherwise, the parent node will be modified and while child node creation,
and will get the same key.

It doesn't seem to be an actual bug,
but such keys look very confusing in the debugger.