fix(pulumi): include build deps in plugin commands
In 0.12, we included build dependencies for Pulumi modules (e.g. when
previewing). A build dependency (e.g. on an `exec` Module or Build)
is a common way to ensure that the relevant language SDK is installed
for the module/action before running Pulumi.

We didn't bring this logic along when we adapted the Pulumi plugin
commands for 0.13—this is fixed here.

Also added a lightweight test project to test this logic (which should
also serve us well when making further fixes to this plugin).