fix(otel): ensure OTEL sends final span when -o opt is used
Before this fix, when the -o/--output option was used, Garden would call
the `shutdown` function early which prevented the final OTEL span to be
emitted.

Now we only run the shutdown logic and (wait for output flush) at the
very end of the command run.

This also means that the `Cli.run()` method no longer has
hard-to-predict side effects, depending on what flags were passed.