fix(otel): do not attempt to shutdown if target exporter not set
Before this fix, attempting to shutdown the OTEL SDK would hang if
called before a target exporter is set. The target exporter is set in
the Garden class so this would e.g. happen when running 'garden help'
and cause a delay before the attempt to shutdown OTEL times out.

With this commit we now check whether an exporter is configured, either
via env vars or the ReconfigurableExporter, and only attempt to shut down
if that's the case.

Note that we need to handle this in the enclosing context as opposed to
having the ReconfigurableExporter handle it itself because the exporter
doesn't "know" if a target exporter will be set later or not.