fix(k8s): correctly set image pull secret on sync pod (#6533)
Before this fix, we'd add image pull secrets with name and namespace to
the Pod spec but the namespace field isn't allowed so applying the
manifest fails.

Now we only add the name. Also added test for the 'configureSyncMode'
function.

The compiler didn't catch this because it allows excess properties
(except for object literals).