fix: fail-safe dependency config lookup in action config processing (#6549)
* fix: handle potentially wrong action names produced from templates

Handle the case when a templated action name from a config template is evaluated to `null` or `undefined`.
Such resolved templatded values are  interpolated into the string values for the action names.

Co-authored-by: Thorarinn Sigurdsson <thorarinnsigurdsson@gmail.com>

* chore: better wording in the log message

Address a comment from CR.

Co-authored-by: Thorarinn Sigurdsson <thorarinnsigurdsson@gmail.com>

* fix: fail-safe dependency config lookup in action config processing

Handle potentially wrong action names produced from configuration templates,
when a templated action name from a config template is evaluated to `null` or `undefined`.
Such resolved template values are interpolated into the string values for the action names.
This is how the current template resolution engine works.
We should skip such dependencies instead of failing fast with an error.

* chore: better wording in the warn message

Co-authored-by: Thorarinn Sigurdsson <thorarinnsigurdsson@gmail.com>

---------

Co-authored-by: Thorarinn Sigurdsson <thorarinnsigurdsson@gmail.com>