fix(local-k8s): ensure correct nginx status when getting env status
We use the `clusterType` config field to determine
the status of the nginx ingress controller when applicable for the
`local-kubernetes` provider.

Before this fix however, we'd set the `clusterType` on the provider
config (if needed) in the `prepareEnvironment` handler and not the
`getEnvironmentStatus` handler.

This basically resulted in the `getEnvironmentStatus` handler always
returning that nginx wasn't ready, then the `clusterType` would be
set in the `prepareEnvironment` and the status checked again, this time
correctly.