feat(templates): allow templating in template path
Added a new template variable for `RenderTemplate` configs:
`${template.path}`. This is the relative path from the project root
to the config template being rendered. This is useful when you want
to use the full path to a file or directory next to the
`ConfigTemplate` that's being rendered (e.g. to use a shared
Dockerfile that belongs the template, every time you render that
template).