fix(k8s): ensure image pull and copy secrets are always created
Before this fix, we'd copy the image pull and copy secrets in the
individual action handlers and in some cases simply forgot to do that.

E.g. if using the cloud builder with a `kubernetes` Deploy type then
no image pull secret would be copied to the target namespace.

Similarly there are code paths where secrets from the `copySecrets` spec
weren't copied to the target namespace.

This fixes that by ensuring the secrets when the environment is prepared
and in turn removing the calls to `ensureSecret` from the individual
action handlers.