chore(framework): mark `getEnvironmentStatus` handler as deprecated (#6774)
In a previous commit we changed the behaviour such that the
`getEnvironmentStatus` handler isn't called before preparing an
environment an instead the `prepareEnvironment` handler is always
called.

This leaves the `getEnvironmentStatus` pretty useless and it's currently
only called with `statusOnly` commands like `garden get status`.

In these cases we might as well set the environment status to `unknown`
and remove these handlers altogether.