fix(k8s): do not throw if paused resource is missing
Before this fix, we'd check for paused Helm resources by listing all
rendered resources and not handling the error if the resource isn't
deployed in the first place.

Furthermore, we were missing an `await` statement so the 404 error from
a missing resource would pop up at a random time, usually after the
status handler returns.

So essentially a `helm` Deploy could fail if using AEC and a potentially
paused resource was missing.

Now we correctly ignore rendered resources that haven't been deployed
and await for the API call.