fix(k8s): ensure image pull secret is always created for K8s Deploy (#6795)
This fixes an issue were in some (edge) cases the image pull secret is
created in the target namespace is using the `kubernetes` Deploy type.

I say edge case because the image pull secret is usually ensured in other
code paths, e.g. when initialising in-cluster builders and as AFAICT
this only happens if using the cloud builder in the k8s Deploy type.

The operation is idempotent so calling it again here should be fine.

TODO @myself: I want to refactor this flow a little bit in a follow up
commit but wanted to get the fix out the way first.