fix: allow access access to variables with unresolvable $merge when resolving built-in keys like dependencies early
Add an `isFinalContext` flag to `VariablesContext`, encapsulating the behaviour we require where we have to allow access to variables despite a `$merge` operation on the same nesting level cannot be resolved.

We only store resolved variables in `ResolveActionTask`, and rely on the nature of lazy evaluation to fail when we have to.