feat: allow cross-referencing variables in the same scope (#6814)
* feat: allow cross-referencing variables in the same scope

Fixes #2730

* feat: allow variable cross-referencing in the project configuration

Avoid resolving variables before creating the variable context;
Otherwise variables cannot reference parent scope and their peers.

* test: add basic test coverage in general and project-level vars

* fix: display resolved variables in `garden get config`

* fix: make sure that variables have access to full context

We need to make sure variables have access to the entire surrounding context, but we also retain fine-grained control over precendence and cross-referencing.

* fix: ensure that higher-precedence variables can refer to root context variables

* test: adapt assertions in project resolve test

* test: add test for action crossreferences

* test: delete unused files

* fix: allow access access to variables with unresolvable $merge when resolving built-in keys like dependencies early

Add an `isFinalContext` flag to `VariablesContext`, encapsulating the behaviour we require where we have to allow access to variables despite a `$merge` operation on the same nesting level cannot be resolved.

We only store resolved variables in `ResolveActionTask`, and rely on the nature of lazy evaluation to fail when we have to.