refactor(terraform): don't init in status handler
Before this fix, we were calling `terraform init` in the
`getStatusHandler` which is an undesired side effect.

With this refactor we now call it in the `prepareEnvironment` handler
which is were it should be.

Also note that in PR #6706 we changed the provider init flow such that
we now always call `prepareEnvironment` and the `getStatusHandler`
isn't really used any more.

That's why I also removed the call to that handler from the
`prepareEnvironment` function and instead moved the logic there.

It's basically duplicated across both handlers now which is fine because
we're removing the status handler in 0.14.