improvement(terraform): support dynamic backends
This change introduces a `backendConfig` field on the Terraform
provider and action types.

This field allows users to dynamically set the backend config for
their Terraform stacks.

Terraform itself doesn't allow variables in the backend config
so this can be used instead to dynamically set the backend
based on e.g. the environment.