fix(dev): fix unresolved templates in cmd results
We were including the raw graph results in the results of commands
requested over HTTP from the live page. This was always unnecessary (and
the data structures involved were excessively large), but only started
resulting in errors with our recent templating improvements (previously,
we'd just leave the unresolved template strings in there).