fix(dev): fix unresolved templates in cmd results
We were including the raw graph results in the results of commands
requested over HTTP from the live page. This was always unnecessary (and
the data structures involved were excessively large), but only started
resulting in errors with our recent templating improvements (previously,
we'd just leave the unresolved template strings in there).

We can clean this up further in 0.14 when we get rid of `graphResults`
from `ProcessCommandResult` (which will require fixing a bunch of test
cases which currently use that field).