docs: collapse sections
Before this change, we'd group docs pages into high level sections like
"Overview" and "Use Cases" and show every sub page under that section.
That takes up a lot of space in the left navbar so now we make the
top-level sections collapsable and hide the sub pages by default.

This is in preparation for further docs improvements we're working on.