improvement(core): fixes to output reference logic
We now infer an `executed: true`  dependency for action output
references when the referenced action is a Build, regardless of whether
the output is a static or a runtime output.

We only do this for API version v2 (and keep the old behavior for v1 and
v0).

This improves the user experience for the common use-case of referencing
a container image in a runtime resource (like a `helm` Deploy), where
the user intent is almost always that the referenced build should exist
(i.e. the dependency should be processed) before the runtime resource is
processed (i.e. deployed or run).

This is because Builds generally don't have side-effects other than
producing artifacts, whereas Deploys and Runs often do.

Note: We could also always execute Test actions that are referenced,
but we'll stick with only Builds for now.