docs: change docs baseurl to latest version
I think we should consider prefixing all docs with the version number,
and forwarding from / to the latest version. @eysi do you think that's
possible?
Otherwise older versions will always link to 404 when we restructure
docs, and it's kind of a hassle to change this constant back and forth
between launches