ci: simplify AWS setup in circleci
Let's just set the AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY directly in the circleci context, rather than remapping it or relying on profile. this simplifies the code and makes it easier to troubleshoot and understand.
This hopefully also fixes the AWS auth issue we have on main.

co-authored-by: vova <vladimir@garden.io>