fix(template): do not throw circular reference error
Fixes a condition we overlooked when implementing the new cross references feature (introduced in
https://github.com/garden-io/garden/pull/6814)

In old garden versions, template expressions like
```
variables:
  foo: ${var.foo}
```
where totally normal, as ${var.foo} just referred to a value in the
parent context.

With cross referencing in mind, we will try to resolve var.foo in the
peer variable context, but if that leads to a circular reference error
we fall back to resolving the value in the parent context.

This seems like a good compromise to preserve the new functionality
while retaining the behaviour from older garden versions.