Release 0.14.7 (#7599)
* chore(release): add step for testing dev command

Added a manual step to `RELEASE.md` for manually testing the `dev`
command during the release process.

* chore(docs): re-enable generated docs check

A few months back, we apparently disabled the step in
`scripts/check-docs.sh` which checks whether the generated docs are up
to date. We re-enable that here (and also include updated generated
docs—them being out of date prompted this fix).

* chore(release): bump version to 0.14.7