Refactor cloud API classes (#7633)
* refactor(core): consistent API naming + remove backend classes

This commit changes most references to the classes previously known as
GardenCloud and GrowCloud, and cloud/cloudApiV2, to cloud legacy and
cloud respectively.

It also removes the GardenBackend classes and corresponding abstractions
which were only really used for the login/logout commands and instead
moves their methods to normal functions.

* refactor(core): more API renaming + rename dirs