feat(core): add utility commands for variables (#7674)
* feat(core): add utility commands for variables

This commit adds a handful of utility commands for listing variables,
variable lists, and users as well as a command for creating remote
variables.

All of this in the context of the new Garden Cloud backend at
https://app.garden.io.

Note that variables are not yet GA in app.garden.io so commands that
don't work yet are hidden.

It also moves a few tests around that where left behind when we refactor
the cloud API classes.

* chore(core): address PR comments