chore(ci): fix apt-get command missing '-y' flag (#7682)
This commit fixes a couple of things I bumped into when releasing
0.14.8.

One was that there's apt-get install command that's missing the "-y"
flag and timing out in CI.

I'm assuming this is coming up now because we merged a dependabot PR that
updated the Dockerfile base image. Not sure why this hasn't come up
before though.

Another issue was with an expired token and simply add a comment in the
relevant place.