feat(core): add `version.excludeFiles` field (#7689)
Yet another mechanism to control version computation for actions.
From the reference docs:

---

### `version.excludeFiles[]`

Specify one or more file paths that should be ignored when computing the version hash for this action.

Specify in the same format as the `include` field. You may use glob patterns here.

For example, you might have a file that naturally changes for every build, such as a compiled binary (that isn't deterministic down to the byte), that you need to have in the build but shouldn't affect the version. You could solve for that with something like this:

```yaml
include:
  - some/compiled/binary
version:
  excludeFiles:
    - some/compiled/binary
```

---