fix(core): only send valid JSON lines to progressui (#7732)
The buildkit output can either be valid JSON or plaintext. We were
sending both to progressui which causes a lot of warning logs when the
progressui is unable to parse the plaintext logs.

This has always been an issue per se but we only recently started
logging these warnings so we hadn't noticed until now.

This fixes that and ensures we don't send plaintext logs to the
progressui tool in the first place.