feat(core): emit deploy run results to Garden Cloud
With this change we include the results with the
ACTION_RUN_COMPLETED event if the action kind is Deploy.

This enables us display links to ingresses in the Garden Cloud UI.

The plan was also to include kind specific state such as whether a Test
action was cached in team or local cache but that's a bigger lift and
will need be handled in a separate PR.

The protobuf schema already accounts for those fields as well though.