chore(ci): fixed a Wine/Rust-related CI error
After updating Rust in CI, we got errors from wine indicating problems
with 32-bit binaries and the lack of a window server. These are fixed
here.

chore(ci): bump Rust to 1.88.0 for build dist jobs

The `build-dist-macos` and `build-dist-linux-windows` CirleCI jobs were
failing with:

  feature `edition2024` is required
  The package requires the Cargo feature called `edition2024`, but that
  feature is not stabilized in this version of Cargo (1.83.0
  (5ffbef321 2024-10-29))

This commit bumps Rust to version 1.88.0 to fix that.