feat(cloud): emit command heartbeat events
We now emit a heartbeat event every 5 seconds while a command is
running. This will be used by Cloud to determine whether a command that
hasn't streamed other events or logs for a while is still doing work or
whether something else is wrong (e.g. a network problem that prevents
the events from making their way to the backend).