fix(k8s): fix fail-fast Helm deploys
This fixes an issue with our fail-fast Helm deploy logic, where if a
previously unhealthy workload belonging to the chart was already
deployed, that unhealthy state would be treated as a deploy failure
before the new version of the resource had time to be committed to the
API server and processed by the controller.

This was essentially a race condition because of the way we started
waiting for resources immediately after we invoked the Helm CLI (whereas
other code locations where we use `waitForResources` await at least the
API operations going through).

The fix involved waiting for resources whose specs could be expected to
be changed by the Helm operation. To make this more solid we (with help
from our LLM friends) made the spec comparison code much more thorough
by attempting to exhaustively address default fields etc. across most
built-in resource types.

This should also make `comparedDeployedResources` more solid in certain
cases.