fix(k8s): fix fail-fast Helm deploys
This fixes an issue with our fail-fast Helm deploy logic, where if a
previously unhealthy workload belonging to the chart was already
deployed, that unhealthy state would be treated as a deploy failure
before the new version of the resource had time to be committed to the
API server and processed by the controller.

This was essentially a race condition because of the way we started
waiting for resources immediately after we invoked the Helm CLI (whereas
other code locations where we use `waitForResources` await at least the
API operations going through).

The fix involved checking the statuses and generation numbers of the
previously deployed unhealthy workloads, and waiting (for a maximum of
10 seconds, as a safety measure) for the generation of the updated
resources to be incremented by the API (indicating that the API
resources have been updated).