improvement(cloud): default orgs (#7852)
* improvement(cloud): default orgs

For enterprise users who have migrated to app.garden.io but haven't yet
updated their project configuration to add an `organizationId`, we now
default to looking up the organization that their old project ID is
associated with. This makes use of a new API endpoint that was written
just for this purpose.

* chore: minor improvements and tests

* chore: log tweaks around API init

* chore: fix formatting issue

---------

Co-authored-by: Jon Edvald <edvald@gmail.com>
Co-authored-by: Eyþór Magnússon <eysi09@gmail.com>