fix(cloud): fix to default org flow (#7860)
The previous implementation didn't prevent migrated single-tenant
users from selecting another org than the one their project ID
belongs to, leading to confusing behaviour.

Here, if a project ID is present, we ensure that the organization ID
belonging to it (if any) is always used, and is automatically written
into the project config on disk.