chore(jib): update tool versions
It's been a while since we updated the versions of OpenJDK, maven and
gradle that Garden has built-in awareness of.

There were also breakages in the jib-container example project, partly
due to the base images on Docker hub now requiring a login. This was
fixed by using a publicly available registry instead.